Solved 5, finally

+++ 5/main.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
 
-int gcd(int m, int n) {
+long gcd(long m, long n) {
   // assure m >= n
   if (m < n) {
-    int tmp = m;
+    long tmp = m;
     m = n;
     n = tmp;
   }
 
   while (n > 0) {
-    int tmp = n;
+    long tmp = n;
     n = m % tmp;
     m = tmp;
   }
@@ -17,34 +17,34 @@ int gcd(int m, int n) {
   return m;
 }
 
-int lcm(int m, int n) {
+long lcm(long m, long n) {
   return m * n / gcd(m, n);
 }
 
-int solve(int target) {
-  int a = 1;
-  for (int i = 1; i <= target; i++) {
+long solve(long target) {
+  long a = 1;
+  for (long i = 1; i <= target; i++) {
     a = lcm(a, i);
   }
   return a;
 }
 
 int main(int argc, char **args) {
-  int a;
+  long a;
 
   a = solve(5);
   if (a != 60) {
-    printf("expected %d but got %d\n", 60, a);
+    printf("expected %ld but got %ld\n", (long)60, a);
     return 1;
   }
 
   a = solve(10);
   if (a != 2520) {
-    printf("expected %d but got %d\n", 2520, a);
+    printf("expected %ld but got %ld\n", (long)2520, a);
     return 1;
   }
 
   a = solve(20);
-  printf("answer = %d\n", a);
+  printf("answer = %ld\n", a);
   return 0;
 }
